Improve search for single mementos, add unit tests, fixes #476
